requirements.txt
atlassian-python-api
tabulate